Refines site aesthetics and readability (#6)
* Adds a footer and adjusts main section layout

Adds a basic footer to the page.

Adjusts the main content area to improve spacing and visual appeal by increasing width, adding margin, padding and gap.

* Removes full-page background styling

Removes the background styling applied to the html element.
This likely moves the styling to a more appropriate container element
or component.

* Updates dark mode color palette

Adjusts the color palette for dark mode to use a slightly lighter shade of black (rgba(10, 10, 10, 1)) instead of pure black (rgba(0, 0, 0, 1)).

This change improves visibility and reduces eye strain in dark mode by providing slightly better contrast.

* Adjusts background blur effect intensity

Increases the background's top position to intensify the blur effect, enhancing visual appeal.

* Improves spacing for readability

Adjusts the spacing of headings and main content
to enhance visual appeal and readability.

Sets the top margin of the h1 element to zero and
increases the bottom margin of the main element.

* Updates CV and contact icons to use same color

Ensures visual consistency by using the same color for the CV and contact icons as the email icon.

Also refactors the CV icon to use a simpler SVG path for better rendering and maintainability.

* Restructures "Know more" links

Replaces simple anchor tags with a div structure
to facilitate styling and visual enhancements.
This change improves the presentation and
potentially the user experience of the links.

* Updates site typography for better readability

Improves the site's typography by switching to "Open Sans" font, adjusting font sizes, and increasing margins.

This enhances readability and provides a more modern look and feel.

* Updates accent color to purple

Updates the accent color to purple for better visual appeal
and consistency with the overall design.

* Updates styling for improved aesthetics

Enhances the visual appeal by adjusting colors, borders, and background opacity.

Modifies the accent color to align with the branding.
Improves card appearance with a more subtle background on hover.
Adds color to the h3 heading.