Improves responsiveness and layout of main content (#9)
* Improves responsiveness and layout

Adds a viewport meta tag to enhance mobile viewing.

Adjusts the main layout and font sizes to better adapt
to different screen sizes, improving the overall user
experience on smaller devices. Introduces media queries
to revert to the original layout on larger screens.

* Improves hero section appearance

Enhances the hero section's visual appeal by increasing the main heading's font size and centering it.

Adjusts main content padding for better spacing and re-enables the display of icons on larger screens, improving overall layout and responsiveness.

* Improves backdrop filter application.

Applies backdrop filter to specific sections instead of the main container.

This change isolates the backdrop filter effect to the about, projects, and skills sections, creating a stacking context that prevents unintended blurring of elements.

* Temporarily removes the projects section.

Removes the projects section from the main page.

This change is temporary and the section will be
reintroduced in a future update.

* Improves main section layout and structure.

Refactors the main section to use relative positioning and height.
Removes unused backdrop-filter properties to streamline styling.
These changes aim to enhance the overall layout and visual appeal
of the main content area.

* Increases main section padding

Updates the padding at the bottom of the main section to improve spacing and overall visual balance.